秒杀服务完成：学习秒杀系统设计，多重校验及MQ削峰；页面渲染

+++ gulimall-seckill/src/main/java/com/zyj/gulimall/seckill/to/SecKillSkuRedisTo.java
@@ -35,15 +35,16 @@ public class SecKillSkuRedisTo {
     /**
      * 秒杀总量
      */
-    private BigDecimal seckillCount;
+    private Integer seckillCount;
     /**
      * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * 排序
      */
     private Integer seckillSort;
+
 
     /**
      * 当前商品秒杀的开始时间
